added ReactRouter to the project and scaffolded the Nav bar.
